Add sourceId to collections
This makes it possible to do asynchronous deposits, as the depositor needs to provide a sourceId and then listen for an object to come over the 'deposited' channel that has that sourceId. Then they will be able to know what the druid is for that object.  This is relevant to letting H2 use RabbitMQ for doing deposits.