Turn on date/time validation
Fixes #485

This commit turns on the date/time validator, which has been available for some time but has been waiting for date/time remediation to happen. This is now finished. As part of this work, pull the latest date/time regular expressions over from DSA, so that validation matches report behavior.