Allow generator to reuse custom types
This commit changes the models so that we are working more by reference and less by value, e.g. instead of having the druid regex in every model that has an `externalIdentifier`, point every `externalIdentifier` at the `Druid` model and let the model handle the validation.

Currently the models handle validation through a combination of validating against the OpenAPI spec and using the generated Ruby (dry-) types, and this commit adds a bit more smarts and descriptive oomph to the Ruby side.