Discontinue allowing non-number characters in catkeys
Connects to #561

This commit tightens the regular expression governing what is a valid Symphony catalog record ID. Arcadia confirms this is the right value, and it is aligned with the regex for the new Folio catalog links (of the migrated-from-symphony variety). Additionally, I ran a report in DSA (sul-dlss/dor-services-app#4357) confirming there are zero catkeys in QA, stage, and prod that contain non-number characters.